moved include files for cel parts to be compatible with the cel naming scheme to ease future backports merged back recent extensions and fixes to the recast and celgraph plugins made in cel (e.g. better handling of sector boundaries) reverted parallalized BuildTile as it didn't work properly

+++ trunk/src/plugins/common/celgraph/celgraph.h
@@ -32,7 +32,7 @@
 #include "ivaria/conin.h"
 #include "ivaria/conout.h"
 
-#include <celgraph.h>
+#include "tools/celgraph.h"
 
 #include <stdio.h>
 #include <string.h>
@@ -104,7 +104,7 @@ celPath, csObject, iCelPath, iComponent>
 {
  private:
   iObjectRegistry* object_reg;
-  int cur_node;
+  size_t cur_node;
   csRefArray<iMapNode> nodes;
   
  public:
